consistent roster logic
