+++ DataWorkerService.Tests/AutomationChecks/Games/GameScoreCountCheckTests.cs
@@ -154,7 +154,7 @@ public class GameScoreCountCheckTests : AutomationChecksTestBase<GameScoreCountC
 
         foreach ((GameScore score, var i) in game.Scores.Select((score, i) => (score, i)))
         {
-            score.Team = game.TeamType is TeamType.HeadToHead or TeamType.TagCoop
+            score.Team = game.TeamType is TeamType.HeadToHead
                 ? Team.NoTeam
                 : i % 2 == 0
                     ? Team.Red
