+++ DataWorkerService/AutomationChecks/Games/GameScoreCountCheck.cs
@@ -2,6 +2,7 @@ using Common.Enums.Enums;
 using Common.Enums.Enums.Verification;
 using Common.Utilities.Extensions;
 using Database.Entities;
+using DataWorkerService.Utilities;
 
 namespace DataWorkerService.AutomationChecks.Games;
 
@@ -29,14 +30,14 @@ public class GameScoreCountCheck(ILogger<GameScoreCountCheck> logger) : Automati
             return false;
         }
 
-        if (entity.TeamType is TeamType.TeamVs or TeamType.TagTeamVs)
+        if (entity.TeamType is TeamType.TeamVs)
         {
-            var redValidScoresCount = validScores.Count(gs => gs.Team == Team.Red);
-            var blueValidScoresCount = validScores.Count(gs => gs.Team == Team.Blue);
+            ICollection<GameRoster> rosters = RostersHelper.GenerateRosters(validScores);
+            var playerCountPerTeam = rosters.Select(x => x.Roster.Length).ToArray();
 
-            if (redValidScoresCount == blueValidScoresCount &&
-                redValidScoresCount == entity.Match.Tournament.LobbyTeamSize &&
-                redValidScoresCount + blueValidScoresCount == validScoresCount)
+            if (playerCountPerTeam.Length > 1 && // more than one team
+                playerCountPerTeam.All(x => x == playerCountPerTeam[0]) && // all counts are equal
+                playerCountPerTeam[0] == entity.Match.Tournament.LobbyTeamSize) // all counts are correct
             {
                 return true;
             }
